Sign error due to secret with $ character
special_encode method is somehow working for LTI request validation but if the secret contains $ character, the signature does not match with other LMSs. Due to this, encodeURIComponent is used just for the secret.

I tried to change the special_encode function to use encodeUTIComponent but it breaks LTI launch request validation which was normally working. So, only here the usage is replaced. Need further investigation.