Addition of ContentItemSelectionRequest support
The content extensions were implemented for a previous version of the protocol. Currently, it is not working as here. Only modified constructor to use it as a parser and to reduce the modifications required. None of the methods in this file can be used. I implemented callback request (ContentItemSelection) in acos-server. Could be integrated here.